+++ src/GDAsset/GDAsset.ts
@@ -31,7 +31,7 @@ export default class GDAsset {
   rootNode: string | undefined = undefined;
   nodePath(n: string) {
     if (!this.rootNode || !n) return n;
-    if (n == '.') return GDAsset.nodeCode('/root/' + this.rootNode);
+    if (n == '.') return GDAsset.nodeCode('../' + this.rootNode);
     if (n.startsWith('./')) return GDAsset.nodeCode(n.substring(2));
     return GDAsset.nodeCode(n);
   }