+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 Features currently in early access are marked with 🔒 as they're restricted to [supporters].  
 Features previously in early access are marked with ***vX.Y.Z*** 🔓 as they're unlocked for everyone since that version.
 
+## [Unreleased]
+### Added
+- GDShader: Add `#error` directive to preprocessor and to syntax coloring.
+### Fixed
+- Stronger pattern to match files with more priority than other extensions like godot-tools.
+
 ## [0.1.0] - 2024-10-16
 ### Changed
 - GDShader: Disable url detection by default, since in BBCode it incorrectly includes the `[/url]` end tag too.